CID#191680: Wrapper object use after free (WRAPPER_ESCAPE)
The coverity report says:

  3. escape: The internal representation of local jv escapes into
  const_node->parent, but is destroyed when it exits scope.

but jv itself is returned from the function *moved*, and it lives on
in the caller's scope (as intended).  I'm pretty certain that this
one is false positive.

The additional code added here verifies that the pointers stored in
the child nodes match that of the core of the parent.