Fix json document initialization with nested initializer lists.
This type of initialization

    json::document_tree doc = {
        { true, false, nullptr },
        { 1.1, 2.2, "text" }
    };

failed previously because the inner initializer list instance would
delete all the enclosing node instances before handing them to the
outer initializer list.  The code was "moving" the inner initializer
list to the outer one, which seems to have worked with gcc, but MSVC
does not move the values.

This change makes the initialization node copyable, and copy them
before handing them over to the outer list.  Not ideal, but better
than not working...