Fix AVX2 version of skip_space_and_control().
Turns out that _mm256_cmpgt_epi8() is a comparison of *signed*
integers, which means any value greater than 0x7F is considered
smaller than whitespace thus to be skipped.  We need to work around
this by checking for the signed bit and mark that char disqualified
if it's on.

If there is an equivalent AVX intrinsic for unsigned integers, we
chould use that instead. But so far I couldn't find one.