Handle attribute without namespace correctly.
c.f. https://www.w3.org/TR/REC-xml-names/#defaulting

When an element name does not have an explicit namespace, it takes
the current default namespace (if one is defined). An attribute name,
on the other hand, do NOT use the current default namespace when it
does not have an explicit namespace given; the decision should be up
to the element it belongs to, according to the above specification.

This issue mostly affects those ods and xlsx files generated by 3rd
party software that makes liberal use of default namespaces.