Pick up font names and sizes
Also handle cases where a cell itself has bold font applied but
contains a non-bold part.  Because xls-xml doesn't have a construct
for expressing a non-bold segment, it sets the bold flag with the
base cell and the bolded text segments while the non-bold segment
doesn't have a bold tag.  What this means that we need to interpret
the absence of a bold tag to mean setting a non- bold format explicitly
for that segment.  I'm assuming it's the same situation with the italic
attribute.