Support native unicode filepaths on Windows
This is a massive update.  The idea is to use wmain(int argc wchar_t** argv) as
the entry point function on Windows and propagate use of wchar_t and 16-bit
string types to transport filepath values across the board.  The rest of the
string values are still expected to be in UTF-8.